requirements.txt
async-timeout==4.0.2
boto3==1.26.21
botocore==1.29.21
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
jmespath==1.0.1
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
redis==4.3.5
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
spotipy==2.21.0
urllib3==1.26.13
